Create SrFeO-hcp-v1.f90
I believe this correctly constructs the SrFeO cells, and will find the distance between them. Also added a couple functions at the bottom to read the indexs for outputting data, currently unused but will be useful. Currently the distances are quite correct as there are some distances of 0, and the nearest neighbourcounting is completey wrong and obselete. Need to get the program to track the number of interactions each atom in the center cell has with each other type of atom, or wyckoff position, I'm not sure which I'm looking for. Then need to divide by the number of primitive cell in the unit cell.